Grid Layers and color options implemented.
Working on Conway's Game of Life today. Currently I have set up a back grid layer and a front grid layer. It randomly kills and births new cells on the back grid then places it on the front grid to be displayed and continues the process while changing the colors. The next step is to implement some of the rules from the Game of Life and ensure that no out of bounds errors exist.