Bug fix!
Finally discovered the primary issue with my GOL. I was doing a shallow copy of the front grid to the back causing it not to retain cell data appropriately. It appears I also do not have to do anything to clear the grids because the data in the grids is updated upon assignment and with the rules.